backend/requirements.txt
flask
torch
diffusers
imageio
google-cloud-storage